refactor(linter): Move Python linters to new file
