+++ tasks/linter/__init__.py
@@ -1,3 +1,4 @@
 """Module regrouping all invoke tasks used for linting the `datadog-agent` repo"""
 
 from .old import *  # noqa: F403
+from .python import python  # noqa: F401
