+++ tasks/linter/python.py
@@ -1,0 +1,32 @@
+"""Tasks for linting Python-related files"""
+
+from invoke.tasks import task
+
+from tasks.libs.common.git import get_default_branch
+from tasks.libs.common.utils import running_in_ci
+
+
+@task
+def python(ctx):
+    """Lints Python files.
+
+    See 'setup.cfg' and 'pyproject.toml' file for configuration.
+    If running locally, you probably want to use the pre-commit instead.
+    """
+
+    print(
+        f"""Remember to set up pre-commit to lint your files before committing:
+    https://github.com/DataDog/datadog-agent/blob/{get_default_branch()}/docs/dev/agent_dev_env.md#pre-commit-hooks"""
+    )
+
+    if running_in_ci():
+        # We want to the CI to fail if there are any issues
+        ctx.run("ruff format --check .")
+        ctx.run("ruff check .")
+    else:
+        # Otherwise we just need to format the files
+        ctx.run("ruff format .")
+        ctx.run("ruff check --fix .")
+
+    ctx.run("vulture")
+    ctx.run("mypy")
